+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### [Unreleased]
+## [0.15.6] - 2020-02-24
+- Updated tick marks in slider widget
+
+## [0.15.5] - 2020-02-22
+- Fixed hidden inputbox
+- Added stacked radio/checkbox/radio item
+
+## [0.15.4] - 2020-02-21
+- Fixed activity access
+
+## [0.15.3] - 2020-02-22
+- Fixed token value format
+
 ## [0.15.2] - 2020-02-22
 - Implemented smooth slider